refactor: 에너지 소모 저장 절대 시간 값 사용으로 변경

+++ Assets/02.Scripts/OutGame/Save/Data/HelperSaveData.cs
@@ -7,6 +7,6 @@ public class HelperSaveData
     public int Level;
     public int Grade;
     public int Experience;
-    public float Energy = -1f;       // -1: 저장 안 됨 → 최대치로 시작
-    public float EnergySavedAt;      // Time.realtimeSinceStartup 기준 저장 시각
+    public float Energy = -1f;       
+    public long EnergySavedAt;
 }
